+++ examples/marker_lsl_sim.py
@@ -22,7 +22,7 @@ import os
 
 from pylsl import StreamInfo, StreamOutlet
 
-from bci_essentials.sources.xdf_sources import XdfEegSource, XdfMarkerSource
+from bci_essentials.io.xdf_sources import XdfEegSource, XdfMarkerSource
 from bci_essentials.utils.logger import Logger  # Logger wrapper
 
 # Instantiate a logger for the module at the default level of logging.INFO
