+++ examples/mi_ch_select_test.py
@@ -6,7 +6,7 @@ Test Motor Imagery (MI) classification offline using data from an existing strea
 import os
 
 # bci_essentials
-from bci_essentials.sources.xdf_sources import XdfEegSource, XdfMarkerSource
+from bci_essentials.io.xdf_sources import XdfEegSource, XdfMarkerSource
 from bci_essentials.eeg_data import EEG_data
 from bci_essentials.classification.mi_classifier import MI_classifier
 from bci_essentials.utils.logger import Logger  # Logger wrapper
