+++ examples/mi_unity_backend.py
@@ -1,4 +1,4 @@
-from bci_essentials.sources.lsl_sources import LslEegSource, LslMarkerSource
+from bci_essentials.io.lsl_sources import LslEegSource, LslMarkerSource
 from bci_essentials.eeg_data import EEG_data
 from bci_essentials.classification.mi_classifier import MI_classifier
 from bci_essentials.utils.logger import Logger  # Logger wrapper
