+++ tests/test_eeg_data.py
@@ -1,6 +1,6 @@
 import unittest
 
-from bci_essentials.sources.sources import MarkerSource, EegSource
+from bci_essentials.io.sources import MarkerSource, EegSource
 from bci_essentials.eeg_data import EEG_data
 from bci_essentials.classification.null_classifier import Null_classifier
 from bci_essentials.utils.logger import Logger  # Logger wrapper
